refactor wikipedia out to sources dir
