+++ src/covidify/sources/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+
+from . import wikipedia, github
+
+__all__ = ['wikipedia', 'github']
